type change u8 => u32
This value cant be greater than 64, number of bits in u64 so u8 enough.
But best performance is wiht u32!?

with u32:
  Time (mean ± σ):     355.6 ms ±   2.5 ms    [User: 354.2 ms, System: 1.3 ms]
  Range (min … max):   353.6 ms … 361.6 ms    10 runs

before, with u8:
  Time (mean ± σ):     358.5 ms ±   0.9 ms    [User: 355.7 ms, System: 2.8 ms]
  Range (min … max):   357.9 ms … 360.9 ms    10 runs

+++ src/bit_reader.zig
@@ -14,7 +14,7 @@ pub fn BitReader(comptime ReaderType: type) type {
         // Internal buffer of 64 bits
         bits: u64 = 0,
         // Number of bits in the buffer
-        nbits: u8 = 0,
+        nbits: u32 = 0,
 
         const Self = @This();
